LayoutTests/imported/w3c/web-platform-tests/intersection-observer/client-rect-expected.txt


PASS IntersectionObserverEntry.boundingClientRect should match target.boundingClientRect() 
FAIL First rAF should generate notification. assert_equals: entries[0].boundingClientRect.left expected 8 but got 0
